Create labels.ts
